backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
yfinance>=0.2.40
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2